refactor: follow plugin naming pattern
