+++ packages/opentelemetry-plugin-grpc/src/utils.ts
